+++ sdk/tables/data-tables/src/TableServiceClient.ts
@@ -37,15 +37,15 @@ import { GeneratedClient } from "./generated/generatedClient";
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
 import { Pipeline } from "@azure/core-rest-pipeline";
 import { TableItemResultPage } from "./models";
+import { apiVersionPolicy } from "./utils/apiVersionPolicy";
 import { getClientParamsFromConnectionString } from "./utils/connectionString";
 import { handleTableAlreadyExists } from "./utils/errorHelpers";
 import { isCredential } from "./utils/isCredential";
 import { logger } from "./logger";
+import { setTokenChallengeAuthenticationPolicy } from "./utils/challengeAuthenticationUtils";
 import { tablesNamedKeyCredentialPolicy } from "./tablesNamedCredentialPolicy";
 import { tablesSASTokenPolicy } from "./tablesSASTokenPolicy";
 import { tracingClient } from "./utils/tracing";
-import { setTokenChallengeAuthenticationPolicy } from "./utils/challengeAuthenticationUtils";
-import { apiVersionPolicy } from "./utils/apiVersionPolicy";
 
 /**
  * A TableServiceClient represents a Client to the Azure Tables service allowing you