+++ sdk/tables/data-tables/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Support cross tenant authentication [PR#21678](https://github.com/Azure/azure-sdk-for-js/pull/21678)
+
 ### Breaking Changes
 
 ### Bugs Fixed
@@ -16,6 +18,7 @@
 ## 13.1.1 (2022-04-14)
 
 ### Bugs Fixed
+
 - Fixed issue where `deleteTable()` doesn't throw any errors [21408](https://github.com/Azure/azure-sdk-for-js/pull/21408).
 
 ## 13.1.0 (2022-04-07)